requirements.txt
alabaster==0.7.10
Babel==2.5.0
certifi==2016.2.28
coverage==4.4.1
cycler==0.10.0
docutils==0.14
h5py==2.7.1
imagesize==0.7.1
Jinja2==2.9.6
Keras==2.0.9
MarkupSafe==1.0
matplotlib==2.0.2
mock==2.0.0
numpy==1.13.1
olefile==0.44
pbr==3.1.1
Pillow==4.2.1
protobuf==3.5.1
py==1.4.34
pydicom==0.9.9
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.2.1
pytest-cov==2.3.1
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.14.2
scipy==0.19.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
tensorflow==1.0.0
